Include UserID in Nodes and Report Controllers

+++ app/Http/Controllers/NodesController.php
@@ -28,6 +28,7 @@ class NodesController extends Controller
                 ->orOn('judments.id_node2', '=', 'node.id');
         })
             ->where('judments.id_node', $id)
+            ->where('node.user_id', Auth::user()->id)
             ->select('node.id', 'node.descr')
             ->distinct()
             ->get();
@@ -44,6 +45,7 @@ class NodesController extends Controller
                 ->orOn('judments.id_node2', '=', 'node.id');
         })
             ->where('judments.id_node', $id)
+            ->where('node.user_id', Auth::user()->id)
             ->select('node.id')
             ->distinct()
             ->get();
@@ -60,10 +62,11 @@ class NodesController extends Controller
         })
             ->whereIn('judments.id_node', $v)
             ->select('node.id', 'node.descr')
+            ->where('node.user_id', Auth::user()->id)
             ->distinct()
             ->get();
 
-        $objective = Node::get()->where('id', $id);
+        $objective = Node::get()->where('user_id', Auth::user()->id)->where('id', $id);
         $goal = $objective[$id - 1];
 
         return view("objetivos.alternatives")->with('alternatives', $alternatives)->with('goal', $goal);
@@ -78,6 +81,7 @@ class NodesController extends Controller
                 ->orOn('judments.id_node2', '=', 'node.id');
         })
             ->where('node.id', '=', $id)
+            ->where('node.user_id', Auth::user()->id)
             ->select('judments.id_node')
             ->distinct()
             ->get();
@@ -92,12 +96,13 @@ class NodesController extends Controller
                 ->orOn('judments.id_node2', '=', 'node.id');
         })
             ->where('node.id', '!=', $id)
+            ->where('node.user_id', Auth::user()->id)
             ->whereIn('judments.id_node', $ids)
             ->select('node.id', 'node.descr')
             ->distinct()
             ->get();
 
-        $target = Node::where('id', $id)->get();
+        $target = Node::where('user_id', Auth::user()->id)->where('id', $id)->get();
 
         return view("objetivos.comparisons")->with('itens', $criteria)->with('goal', $objective)->with('target', $target)->with('id', $id);
     }
@@ -116,38 +121,18 @@ class NodesController extends Controller
             return view("objetivos.formCreateNode")->with('itens', $itens)->with('nodes', $nodes)->with('descr', $descr)->with('up', $up)->with('level', 0);
         } elseif ($data['type'] == 1) {
 
-            $query = Node::select('descr')->where('id', $up);
+            $query = Node::select('descr')->where('user_id', Auth::user()->id)->where('id', $up);
             $q = $query->first();
-            // foreach($query as $q) { array_push($ids, $q->id_node1); array_push($ids, $q->id_node2); }
-            // $ids = array_unique($ids);
 
             $descr = "Criteria for " . $q->descr . " Decision Problem";
             return view("objetivos.formCreateNode")->with('itens', $itens)->with('nodes', $data['nodes'])->with('descr', $descr)->with('up', $up)->with('level', 1);
         } elseif ($data['type'] == 2) {
 
-            $query = Node::select('descr')->where('id', $up);
+            $query = Node::select('descr')->where('user_id', Auth::user()->id)->where('id', $up);
             $q = $query->first();
             $descr = "Alternatives for " . $q->descr . " Decision Problem";
 
-
-
             return view("objetivos.formCreateNode")->with('itens', $itens)->with('nodes', $data['nodes'])->with('descr', $descr)->with('up', $up)->with('level', 2);
-            //     $query = Judments::
-            //     join('node', function ($join) {
-            //     $join->on('judments.id_node1', '=', 'node.id')
-            //     ->orOn('judments.id_node2', '=', 'node.id');
-            //     })
-            //     ->where('node.id','=', $up)
-            //     ->select('judments.id_node')
-            //     ->distinct()
-            //     ->get();
-
-            // foreach($query as $q) 
-            //     array_push($itens, $q->id_node);
-
-            //     print_r($itens);
-
-            //     return view("objetivos.formCreateNode")->with('itens', $itens)->with('nodes', $data['nodes'])->with('descr', $descr)->with('up',$up); 
         }
     }
 
@@ -182,8 +167,8 @@ class NodesController extends Controller
         }
 
         if ($level == 2) {
-            $id_node1 = Judments::select('id_node1')->distinct()->where('id_node', $up);
-            $id_node2 = Judments::select('id_node2')->distinct()->where('id_node', $up)->union($id_node1)->get();
+            $id_node1 = Judments::select('id_node1')->distinct()->where('user_id', Auth::user()->id)->where('id_node', $up);
+            $id_node2 = Judments::select('id_node2')->distinct()->where('user_id', Auth::user()->id)->where('id_node', $up)->union($id_node1)->get();
 
 
             foreach ($id_node2 as $node) {
@@ -215,6 +200,7 @@ class NodesController extends Controller
                     ->orOn('judments.id_node2', '=', 'node.id');
             })
                 ->where('judments.id_node', $v[$i])
+                ->where('node.user_id', Auth::user()->id)
                 ->select('node.id', 'node.descr')
                 ->distinct()
                 ->get();
@@ -226,9 +212,9 @@ class NodesController extends Controller
         } while ($i < count($v));
         $v = array_unique($v);
         Node::whereIn('id', $v)->delete();
-        Judments::whereIn('id_node', $v)->delete();
-        Judments::whereIn('id_node1', $v)->delete();
-        Judments::whereIn('id_node2', $v)->delete();
+        Judments::where('user_id', Auth::user()->id)->whereIn('id_node', $v)->delete();
+        Judments::where('user_id', Auth::user()->id)->whereIn('id_node1', $v)->delete();
+        Judments::where('user_id', Auth::user()->id)->whereIn('id_node2', $v)->delete();
         return redirect("/nodes");
     }
 
@@ -238,7 +224,7 @@ class NodesController extends Controller
         $data = $request->all();
         $s = explode(";", $data['score0']);
         $up = $s[0];
-        $top = Judments::where('id_node', $s[0])->get();
+        $top = Judments::where('user_id', Auth::user()->id)->where('id_node', $s[0])->get();
         $v = array();
         foreach ($top as $t) {
             array_push($v, $t->id_node1);
@@ -277,98 +263,11 @@ class NodesController extends Controller
                 Judments::where('id_node', $up)
                     ->where('id_node1', $nodes[$i])
                     ->where('id_node2', $nodes[$j])
+                    ->where('user_id', Auth::user()->id)
                     ->update(['score' => $matrix[$nodes[$i]][$nodes[$j]]]);
             }
         }
 
         return redirect("/nodes");
     }
-
-    /*public function UpdateScore($proxy, Request $request)
-    
-    {
-        echo $proxy . "<br>";
-        $score = new Score;
-        $data = $request->all();
-        $x = array(array(array())); //id_node - id_node1 - id_node2
-        $n = array();
-        $up = array();
-        $scores = array();
-
-        for ($i = 0; $i < $data['counter']; $i++) {
-            $s = explode(";", $data['score' . $i]);
-            if ($s[1] != $proxy) array_push($n, $s[1]);
-            if ($s[2] != $proxy) array_push($n, $s[2]);
-            array_push($up, $s[0]);
-            if ($s[1] < $s[2]) {
-                $id_node1 = $s[1];
-                $id_node2 = $s[2];
-            }
-            if ($s[1] > $s[2]) {
-                $id_node2 = $s[1];
-                $id_node1 = $s[2];
-            }
-
-            if ($id_node1 == $proxy) {
-                $x[$s[0]][$id_node1][$id_node2] = $s[3];
-                $scr = $s[3];
-            }
-            if ($id_node2 == $proxy) {
-                $x[$s[0]][$id_node2][$id_node1] = 1 / $s[3];
-                $scr = 1 / $s[3];
-            }
-            echo $s[0] . " - " . $id_node1 . " - " . $id_node2 . " = " . $scr . "<hr>";
-            //Judments::where('id_node', $s[0])->where('id_node1', $id_node1)->where('id_node2', $id_node2)->update(['score' => $s[3]]);
-            Judments::where('id_node', $s[0])->where('id_node1', $id_node1)->where('id_node2', $id_node2)->update(['score' => $scr]);
-        }
-
-        $up = array_unique($up);
-        $n = array_unique($n);
-
-        //ESTOU TRABALHANDO AQUI
-        foreach ($up as $p) {
-            for ($i = 0; $i < count($n); $i++) {
-                for ($j = $i + 1; $j < count($n); $j++) {
-                    if ($n[$i] == $proxy || $n[$j] == $proxy)
-                        $z = 1 / ($x[$p][$proxy][$n[$i]] / $x[$p][$proxy][$n[$j]]);
-                    else {
-                        if($proxy < $n[$i]) {
-                            $bfk1 = Judments::where('id_node', $p)->where('id_node1', $proxy)->where('id_node2', $n[$i])->first();
-                        } else {
-                            $bfk1 = Judments::where('id_node', $p)->where('id_node1', $n[$i])->where('id_node2', $proxy)->first();
-                        }
-
-                        if($proxy == $n[$i]) {
-
-                        } else {
-
-                        }
-
-                        if($proxy < $n[$j]) {
-                            $bfk2 = Judments::where('id_node', $p)->where('id_node1', $proxy)->where('id_node2', $n[$j])->first();
-                        } else {
-                            $bfk2 = Judments::where('id_node', $p)->where('id_node1', $n[$j])->where('id_node2', $proxy)->first();
-                        }
-
-                        if($proxy == $n[$j]) {
-                            
-                            $z = $bfk1->score/$bfk2->score;
-                        } else {
-                            echo "<hr>".pow($bfk1->score,(-1))." / ".$bfk2->score."<hr>";
-                            $z = pow($bfk1->score,(-1))/$bfk2->score;
-                        }
-                        
-                        
-                        echo $z;
-                        $z = 1;
-                        
-                    }
-                        
-                    echo ":" . $p . " - " . $n[$i] . " - " . $n[$j] . " == " . $z . "<br>";
-                    Judments::where('id_node', $p)->where('id_node1', $n[$i])->where('id_node2', $n[$j])->update(['score' => $z]);
-                }
-            }
-        }
-        //return redirect("/nodes");
-    }*/
 }
